# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/employees": {
      "get": {
        "operationId": "EmployeesController_list",
        "parameters": [
          {
            "name": "includeInactive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["employees"]
      },
      "post": {
        "operationId": "EmployeesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["employees"]
      }
    },
    "/api/employees/{id}": {
      "get": {
        "operationId": "EmployeesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["employees"]
      },
      "put": {
        "operationId": "EmployeesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmployeeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["employees"]
      },
      "delete": {
        "operationId": "EmployeesController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["employees"]
      }
    },
    "/api/employees/{id}/password": {
      "post": {
        "operationId": "EmployeesController_setPassword",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["employees"]
      }
    },
    "/api/employees/{id}/reactivate": {
      "post": {
        "operationId": "EmployeesController_reactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["employees"]
      }
    },
...
  },
  "info": {
    "title": "OpenClockwork API",
    "description": "Self-hostable working-time tracker — REST + WebSocket surface.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateEmployeeDto": {
        "type": "object",
        "properties": {
          "personalNo": {
            "type": "string",
            "example": "1001",
            "maxLength": 40
          },
          "firstName": {
            "type": "string",
            "example": "Anna",
            "maxLength": 120
          },
          "lastName": {
            "type": "string",
            "example": "Mueller",
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "example": "anna.mueller@openclockwork.test",
            "maxLength": 200
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 120,
            "description": "Initial password — bcrypt-hashed on the server."
          },
          "role": {
            "type": "string",
            "enum": ["Employee", "Manager", "HRAdmin"],
            "example": "Employee"
          },
          "timeModel": {
            "type": "string",
            "enum": [
              "Teilzeit",
              "Vollzeit",
              "Vertrauensarbeitszeit",
              "Gleitzeit"
            ],
            "example": "Vollzeit"
          },
          "weeklyHours": {
            "type": "number",
            "example": 40,
            "minimum": 0
          },
          "annualLeaveDays": {
            "type": "number",
            "example": 30,
            "minimum": 0
          },
          "startDate": {
            "type": "string",
            "example": "2026-04-01",
            "description": "ISO date when the employee starts; target working hours are counted from here."
          },
          "overtimeOpeningBalanceMinutes": {
            "type": "number",
            "example": 0,
            "description": "One-time overtime carry-over in minutes (signed)."
          },
          "bundesland": {
            "type": "string",
            "enum": [
              "BW",
              "BY",
              "BE",
              "BB",
              "HB",
              "HH",
              "HE",
              "MV",
              "NI",
              "NW",
              "RP",
              "SL",
              "SN",
              "ST",
              "SH",
              "TH"
            ],
            "deprecated": true,
            "description": "Legacy alias for a DE-XX holidayCalendar; use holidayCalendar for new clients."
          },
          "holidayCalendar": {
            "type": "string",
            "enum": [
              "NONE",
              "DE-BW",
              "DE-BY",
              "DE-BE",
              "DE-BB",
              "DE-HB",
              "DE-HH",
              "DE-HE",
              "DE-MV",
              "DE-NI",
              "DE-NW",
              "DE-RP",
              "DE-SL",
              "DE-SN",
              "DE-ST",
              "DE-SH",
              "DE-TH"
            ],
            "default": "NONE",
            "description": "Optional holiday preset. NONE makes no public-holiday assumptions; explicit holidayDates work in every country."
          },
          "holidayDates": {
            "example": ["2026-07-01"],
            "description": "Explicit non-working dates in YYYY-MM-DD format; supplement the selected preset.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowDailyBlockBooking": {
            "type": "boolean",
            "default": false,
            "description": "Allow one self-approved fixed-duration block on a configured working day."
          },
          "managerId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "workScheduleId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          }
        },
        "required": [
          "personalNo",
          "firstName",
          "lastName",
          "email",
          "password",
          "role",
          "timeModel",
          "weeklyHours",
          "annualLeaveDays",
          "startDate"
        ]
      },
      "UpdateEmployeeDto": {
        "type": "object",
        "properties": {
          "personalNo": {
            "type": "string",
            "maxLength": 40
          },
          "firstName": {
            "type": "string",
            "maxLength": 120
          },
          "lastName": {
            "type": "string",
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "maxLength": 200
          },
          "role": {
            "type": "string",
            "enum": ["Employee", "Manager", "HRAdmin"]
          },
          "timeModel": {
            "type": "string",
            "enum": [
              "Teilzeit",
              "Vollzeit",
              "Vertrauensarbeitszeit",
              "Gleitzeit"
            ]
          },
          "weeklyHours": {
            "type": "number",
            "minimum": 0
          },
          "annualLeaveDays": {
            "type": "number",
            "minimum": 0
          },
          "startDate": {
            "type": "string",
            "example": "2026-04-01"
          },
          "overtimeOpeningBalanceMinutes": {
            "type": "number"
          },
          "bundesland": {
            "type": "string",
            "enum": [
              "BW",
              "BY",
              "BE",
              "BB",
              "HB",
              "HH",
              "HE",
              "MV",
              "NI",
              "NW",
              "RP",
              "SL",
              "SN",
              "ST",
              "SH",
              "TH"
            ],
            "deprecated": true
          },
          "holidayCalendar": {
            "type": "string",
            "enum": [
              "NONE",
              "DE-BW",
              "DE-BY",
              "DE-BE",
              "DE-BB",
              "DE-HB",
              "DE-HH",
              "DE-HE",
              "DE-MV",
              "DE-NI",
              "DE-NW",
              "DE-RP",
              "DE-SL",
              "DE-SN",
              "DE-ST",
              "DE-SH",
              "DE-TH"
            ],
            "default": "NONE",
            "description": "Optional holiday preset. NONE makes no public-holiday assumptions; explicit holidayDates work in every country."
          },
          "holidayDates": {
            "example": ["2026-07-01"],
            "description": "Explicit non-working dates in YYYY-MM-DD format; supplement the selected preset.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowDailyBlockBooking": {
            "type": "boolean",
            "description": "Allow one self-approved fixed-duration block on a configured working day."
          },
          "managerId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "workScheduleId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "SetPasswordDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 120
          }
        },
        "required": ["password"]
      },
...
    }
  }
}